+++ src/main/kotlin/dk/ksp/algotrading/entity/StockPrice.kt
@@ -3,17 +3,18 @@ package dk.ksp.algotrading.entity
 import jakarta.persistence.Entity
 import jakarta.persistence.Id
 import jakarta.persistence.Table
+import java.math.BigDecimal
 
 @Entity
 @Table(name = "stock_prices")
 class StockPrice(
-    val currentPrice: Double,
-    val change: Double,
-    val percentChange: Double,
-    val highPrice: Double,
-    val lowPrice: Double,
-    val openPrice: Double,
-    val previousClosePrice: Double,
+    val currentPrice: BigDecimal,
+    val change: BigDecimal,
+    val percentChange: BigDecimal,
+    val highPrice: BigDecimal,
+    val lowPrice: BigDecimal,
+    val openPrice: BigDecimal,
+    val previousClosePrice: BigDecimal,
     @Id
     val timestamp: Long,
 )
